pyproject.toml
[tool.poetry]
name = "books-shop"
version = "0.1.0"
description = ""
authors = ["Павел <king-he@mail.ru>"]
readme = "README.md"
packages = [{include = "books_shop"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.6"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
